adding some policy

+++ app/Http/Controllers/ListingController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ListingController extends Controller
 {
@@ -25,6 +26,7 @@ class ListingController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create');
         return inertia('Listing/Create');
     }
 
@@ -33,6 +35,7 @@ class ListingController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create');
         // $request->user()->listings()->create()
         Auth::user()->listings()->create(
             $request->validate([
@@ -55,6 +58,10 @@ class ListingController extends Controller
      */
     public function show(Listing $listing)
     {
+        // if(Auth::user()->cannot('view', $listing)) {
+        //     abort(403);
+        // }
+        Gate::authorize('view', $listing);
         return inertia('Listing/Show', 
             [
                 'listing'=> $listing
@@ -67,7 +74,8 @@ class ListingController extends Controller
      */
     public function edit(Listing $listing)
     {
-          return inertia(
+        Gate::authorize('update', $listing);
+        return inertia(
             'Listing/Edit',
             [
                 'listing' => $listing
@@ -80,6 +88,7 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
+        Gate::authorize('update', $listing);
         $listing->update(
             $request->validate([
                 'beds' => 'required|integer|min:0|max:20',
@@ -102,6 +111,7 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        Gate::authorize('delete', $listing);
         $listing->delete();
 
         return redirect()->back()
